requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pillow
drf-spectacular
celery==5.4.0
redis==5.2.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1